STERI016-413-regression-am: Ensure button renders correctly in cards
